mv customButton, RegisterPage into sep. files
